ui: Extract mouse.c/mouse.h
