+++ src/nvim/edit.c
@@ -56,6 +56,7 @@
 #include "nvim/tag.h"
 #include "nvim/term.h"
 #include "nvim/ui.h"
+#include "nvim/mouse.h"
 #include "nvim/undo.h"
 #include "nvim/window.h"
 #include "nvim/os/event.h"
