+++ src/nvim/normal.c
@@ -57,6 +57,7 @@
 #include "nvim/tag.h"
 #include "nvim/term.h"
 #include "nvim/ui.h"
+#include "nvim/mouse.h"
 #include "nvim/undo.h"
 #include "nvim/window.h"
 #include "nvim/os/event.h"
